secure-sms/requirements.txt
# Use a flexible version range for Flask to ensure compatibility with
# hosted package indices.  Fixing the upper bound allows for minor
# updates while avoiding breaking changes in future major versions.
Flask>=2.3.0,<4.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
email_validator==2.2.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
python-dotenv==1.0.1
cryptography==43.0.0
pyotp==2.9.0
itsdangerous==2.2.0